Breakout robot classes to separate modules
