+++ pylitterbot/robot/models.py
@@ -1,4 +1,4 @@
-"""pylitterbot models"""
+"""pylitterbot robot models"""
 LITTER_ROBOT_4_MODEL = """
 {
     unitId
